# fix typo + faster is_sorted